+++ .gitignore
@@ -2,7 +2,7 @@
 *.class
 
 # Target directory
-*/target/*
+*target/*
 
 # Log file
 *.log
